Adding hashtablefile_readonly
Introducing API to support reading an existing hashtable backed by a
file on the disk. This hashtable is not intended to be modified and
hence the API supports only the lookup functionality. A different set
of API's (as outlined) in hashtable.h should be used to generate a
hashtable backed by a file on disk.